Remove building tests from test script
