+++ build.rs
@@ -1,7 +1,6 @@
 use std::{
     env, fs,
     path::{Path, PathBuf},
-    process::Command,
 };
 
 fn main() {
@@ -29,35 +28,4 @@ fn main() {
         out_dir.join("wasm32-unknown-unknown/release/wasi_snapshot_preview1_command.wasm"),
     )
     .unwrap();
-
-    build_rust_test_case(&out_dir, "rust-case-0.2");
-    build_rust_test_case(&out_dir, "rust-case-0.8");
-    build_rust_test_case(&out_dir, "rust-command");
-
-    let mut cmd = Command::new("tinygo");
-    cmd.arg("build")
-        .current_dir("tests/go-case")
-        .arg("-target=wasi")
-        .arg("-gc=leaking")
-        .arg("-no-debug")
-        .arg("-o")
-        .arg(out_dir.join("go_case.wasm"))
-        .arg("main.go");
-
-    // If just skip this if TinyGo is not installed
-    _ = cmd.status();
-    println!("cargo:rerun-if-changed=tests/go-case");
-}
-
-fn build_rust_test_case(out_dir: &PathBuf, name: &str) {
-    let mut cmd = Command::new("cargo");
-    cmd.arg("build")
-        .current_dir(&format!("tests/{name}"))
-        .arg("--release")
-        .arg("--target=wasm32-wasi")
-        .env("CARGO_TARGET_DIR", out_dir);
-
-    let status = cmd.status().unwrap();
-    assert!(status.success());
-    println!("cargo:rerun-if-changed=tests/{name}");
 }
